agents/requirements.txt
jsonref
dotenv
azure.identity
azure.ai.projects
azure.ai.agents
